test: updated tests to not run in parallel
