+++ test/tap-parallel-not-ok/blog/update-blog.test.js
@@ -1,6 +1,9 @@
 import tap from 'tap';
-import { build } from '../../src/app.js';
+import { build } from '../../../src/app.js';
 import 'must/register.js';
+import Chance from 'chance';
+
+const chance = new Chance();
 
 tap.mochaGlobals();
 
@@ -9,8 +12,60 @@ const prefix = '/api';
 describe('Updateing a blog should work', async () => {
   let app;
 
+  const newUser = {
+    username: chance.email({ domain: 'example.com' }),
+    firstName: chance.first(),
+    lastName: chance.last(),
+    password: chance.string({ length: 10 })
+  };
+
+  let cookie = '';
+
   before(async () => {
-    app = await build();
+    app = await build({
+      forceCloseConnections: true
+    });
+  });
+
+  it('should return the created user object', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/user`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(newUser)
+    });
+
+    // checks if status code is 200
+    response.statusCode.must.be.equal(200);
+    const result = await response.json();
+    result.username.must.be.equal(newUser.username);
+    result.firstName.must.be.equal(newUser.firstName);
+    result.lastName.must.be.equal(newUser.lastName);
+    // expect dates to be not null
+    result.createdDate.must.not.be.null();
+    result.updatedDate.must.not.be.null();
+  });
+  it('login should work', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/user/login`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(
+        {
+          username: newUser.username,
+          password: newUser.password
+        }
+      )
+    });
+
+    // checks if status code is 200
+    response.statusCode.must.be.equal(200);
+
+    cookie = response.headers['set-cookie'];
   });
 
   it('should update the title and desc with uuid', async () => {
@@ -28,7 +83,8 @@ describe('Updateing a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newBlog)
     });
@@ -39,7 +95,8 @@ describe('Updateing a blog should work', async () => {
       method: 'PUT',
       url: `${prefix}/blog/${id}`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newerBlog)
     });
@@ -72,7 +129,8 @@ describe('Updateing a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newBlog)
     });
@@ -83,7 +141,8 @@ describe('Updateing a blog should work', async () => {
       method: 'PUT',
       url: `${prefix}/blog/${id}`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newerBlog)
     });
@@ -116,7 +175,8 @@ describe('Updateing a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newBlog)
     });
@@ -127,7 +187,8 @@ describe('Updateing a blog should work', async () => {
       method: 'PUT',
       url: `${prefix}/blog/${id}`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newerBlog)
     });
@@ -144,5 +205,8 @@ describe('Updateing a blog should work', async () => {
     // expect dates to be not null
     result.createdDate.must.be.equal(createdDate);
     result.updatedDate.must.be.above(updatedDate);
+  });
+  after(async () => {
+    await app.close();
   });
 });
