+++ test/tap-parallel-not-ok/comment/get-many-comment.test.js
@@ -1,6 +1,9 @@
 import tap from 'tap';
-import { build } from '../../src/app.js';
+import { build } from '../../../src/app.js';
 import 'must/register.js';
+import Chance from 'chance';
+
+const chance = new Chance();
 
 tap.mochaGlobals();
 
@@ -9,8 +12,60 @@ const prefix = '/api';
 describe('Getting comments of a blog should work', async () => {
   let app;
 
+  const newUser = {
+    username: chance.email({ domain: 'example.com' }),
+    firstName: chance.first(),
+    lastName: chance.last(),
+    password: chance.string({ length: 10 })
+  };
+
+  let cookie = '';
+
   before(async () => {
-    app = await build();
+    app = await build({
+      forceCloseConnections: true
+    });
+  });
+
+  it('should return the created user object', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/user`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(newUser)
+    });
+
+    // checks if status code is 200
+    response.statusCode.must.be.equal(200);
+    const result = await response.json();
+    result.username.must.be.equal(newUser.username);
+    result.firstName.must.be.equal(newUser.firstName);
+    result.lastName.must.be.equal(newUser.lastName);
+    // expect dates to be not null
+    result.createdDate.must.not.be.null();
+    result.updatedDate.must.not.be.null();
+  });
+  it('login should work', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/user/login`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(
+        {
+          username: newUser.username,
+          password: newUser.password
+        }
+      )
+    });
+
+    // checks if status code is 200
+    response.statusCode.must.be.equal(200);
+
+    cookie = response.headers['set-cookie'];
   });
 
   it('should return the created object with uuid', async () => {
@@ -22,7 +77,8 @@ describe('Getting comments of a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newBlog)
     });
@@ -37,7 +93,8 @@ describe('Getting comments of a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog/${blogId}/comment`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newComment)
     });
@@ -50,7 +107,8 @@ describe('Getting comments of a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog/${blogId}/comment`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newComment2)
     });
@@ -63,7 +121,8 @@ describe('Getting comments of a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog/${blogId}/comment`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newComment3)
     });
@@ -76,7 +135,8 @@ describe('Getting comments of a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog/${blogId}/comment`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newComment4)
     });
@@ -89,7 +149,8 @@ describe('Getting comments of a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog/${blogId}/comment`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newComment5)
     });
@@ -102,7 +163,8 @@ describe('Getting comments of a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog/${blogId}/comment`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newComment6)
     });
@@ -116,6 +178,9 @@ describe('Getting comments of a blog should work', async () => {
 
     const response = await app.inject({
       method: 'GET',
+      headers: {
+        cookie
+      },
       url: `${prefix}/blog/${blogId}/comment`
     });
 
@@ -124,5 +189,8 @@ describe('Getting comments of a blog should work', async () => {
     const result = await response.json();
 
     result.length.must.not.be.above(5);
+  });
+  after(async () => {
+    await app.close();
   });
 });
