+++ test/tap-parallel-not-ok/comment/update-comment.test.js
@@ -1,5 +1,5 @@
 import tap from 'tap';
-import { build } from '../../src/app.js';
+import { build } from '../../../src/app.js';
 import 'must/register.js';
 import Chance from 'chance';
 
@@ -9,12 +9,8 @@ tap.mochaGlobals();
 
 const prefix = '/api';
 
-describe('Logging out a user should work', async () => {
+describe('Updating comments of a blog should work', async () => {
   let app;
-
-  before(async () => {
-    app = await build();
-  });
 
   const newUser = {
     username: chance.email({ domain: 'example.com' }),
@@ -24,6 +20,12 @@ describe('Logging out a user should work', async () => {
   };
 
   let cookie = '';
+
+  before(async () => {
+    app = await build({
+      forceCloseConnections: true
+    });
+  });
 
   it('should return the created user object', async () => {
     const response = await app.inject({
@@ -38,8 +40,6 @@ describe('Logging out a user should work', async () => {
     // checks if status code is 200
     response.statusCode.must.be.equal(200);
     const result = await response.json();
-
-    // all new values should be equal to properties of new blog
     result.username.must.be.equal(newUser.username);
     result.firstName.must.be.equal(newUser.firstName);
     result.lastName.must.be.equal(newUser.lastName);
@@ -67,29 +67,67 @@ describe('Logging out a user should work', async () => {
 
     cookie = response.headers['set-cookie'];
   });
-  it('logout should work', async () => {
-    const response = await app.inject({
-      method: 'GET',
-      url: `${prefix}/user/logout`,
+
+  it('should return the created object with uuid', async () => {
+    const newBlog = {
+      title: 'New Blog to get comment from test',
+      desc: 'New Description to get comment from test'
+    };
+    const createBlog = await app.inject({
+      method: 'POST',
+      url: `${prefix}/blog`,
       headers: {
         'Content-Type': 'application/json',
         cookie
-      }
+      },
+      body: JSON.stringify(newBlog)
     });
 
-    // checks if status code is 200
-    response.statusCode.must.be.equal(200);
-  });
-  it('logout should return error 401, no cookie', async () => {
-    const response = await app.inject({
-      method: 'GET',
-      url: `${prefix}/user/logout`,
+    const { id: blogId } = await createBlog.json();
+
+    const newComment = {
+      text: 'New Comment to update comment from test'
+    };
+
+    const createComment = await app.inject({
+      method: 'POST',
+      url: `${prefix}/blog/${blogId}/comment`,
       headers: {
-        'Content-Type': 'application/json'
-      }
+        'Content-Type': 'application/json',
+        cookie
+      },
+      body: JSON.stringify(newComment)
+    });
+
+    const { commentId, createdDate, updatedDate } = await createComment.json();
+
+    const newerComment = {
+      text: 'Updated comment from test'
+    };
+
+    const response = await app.inject({
+      method: 'PUT',
+      url: `${prefix}/blog/${blogId}/comment/${commentId}`,
+      headers: {
+        'Content-Type': 'application/json',
+        cookie
+      },
+      body: JSON.stringify(newerComment)
     });
 
     // checks if status code is 200
     response.statusCode.must.be.equal(200);
+    const result = await response.json();
+
+    // expect id to be equal to result
+    result.commentId.must.be.equal(commentId);
+    // new values must be equal to new comment
+    result.text.must.be.equal(newerComment.text);
+    // expect updateddate to be updated and createddate to be same
+    result.createdDate.must.be.equal(createdDate);
+    result.updatedDate.must.be.above(updatedDate);
+  });
+  after(async () => {
+    await app.close();
   });
 });
