+++ test/tap-parallel-not-ok/user/logout.test.js
@@ -1,5 +1,5 @@
 import tap from 'tap';
-import { build } from '../../src/app.js';
+import { build } from '../../../src/app.js';
 import 'must/register.js';
 import Chance from 'chance';
 
@@ -9,11 +9,13 @@ tap.mochaGlobals();
 
 const prefix = '/api';
 
-describe('Creating a user should work', async () => {
+describe('Logging out a user should work', async () => {
   let app;
 
   before(async () => {
-    app = await build();
+    app = await build({
+      forceCloseConnections: true
+    });
   });
 
   const newUser = {
@@ -22,6 +24,8 @@ describe('Creating a user should work', async () => {
     lastName: chance.last(),
     password: chance.string({ length: 10 })
   };
+
+  let cookie = '';
 
   it('should return the created user object', async () => {
     const response = await app.inject({
@@ -45,17 +49,52 @@ describe('Creating a user should work', async () => {
     result.createdDate.must.not.be.null();
     result.updatedDate.must.not.be.null();
   });
-  it('should return error 400 since user is creating with existing username', async () => {
+  it('login should work', async () => {
     const response = await app.inject({
       method: 'POST',
-      url: `${prefix}/user`,
+      url: `${prefix}/user/login`,
       headers: {
         'Content-Type': 'application/json'
       },
-      body: JSON.stringify(newUser)
+      body: JSON.stringify(
+        {
+          username: newUser.username,
+          password: newUser.password
+        }
+      )
     });
 
     // checks if status code is 200
-    response.statusCode.must.be.equal(400);
+    response.statusCode.must.be.equal(200);
+
+    cookie = response.headers['set-cookie'];
+  });
+  it('logout should work', async () => {
+    const response = await app.inject({
+      method: 'GET',
+      url: `${prefix}/user/logout`,
+      headers: {
+        'Content-Type': 'application/json',
+        cookie
+      }
+    });
+
+    // checks if status code is 200
+    response.statusCode.must.be.equal(200);
+  });
+  it('logout should return error 401, no cookie', async () => {
+    const response = await app.inject({
+      method: 'GET',
+      url: `${prefix}/user/logout`,
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    });
+
+    // checks if status code is 200
+    response.statusCode.must.be.equal(401);
+  });
+  after(async () => {
+    await app.close();
   });
 });
